move common app-templates code into shared folder
