+++ plugins/@grouparoo/app-templates/src/source/queryv2/index.ts
@@ -1,2 +1,2 @@
 export * from "./meta";
-export * from "../table/options";
+export * from "../shared/options";
