+++ plugins/@grouparoo/app-templates/src/source/shared/getExamples.ts
