+++ plugins/@grouparoo/app-templates/src/source/shared/getFilterOperation.ts
