+++ plugins/@grouparoo/app-templates/src/source/shared/pluginMethods.ts
