+++ plugins/@grouparoo/app-templates/src/source/shared/propertyOptions.ts
