+++ plugins/@grouparoo/app-templates/src/source/shared/recordProperties.ts
