+++ plugins/@grouparoo/app-templates/src/source/shared/recordProperty.ts
