+++ plugins/@grouparoo/app-templates/src/source/shared/scheduleOptions.ts
