+++ plugins/@grouparoo/app-templates/src/source/shared/sourcePreview.ts
