+++ plugins/@grouparoo/app-templates/src/source/shared/uniquePropertyBootstrapOptions.ts
