+++ plugins/@grouparoo/app-templates/src/source/table/index.ts
@@ -1,2 +1,3 @@
 export * from "./meta";
-export * from "./options";
+export * from "./sourceFilters";
+export * from "../shared/options";
